+++ RedPandaIDE/compiler/projectcompiler.cpp
@@ -158,11 +158,9 @@ void ProjectCompiler::writeMakeDefines(QFile &file)
         // Only process source files
         QString RelativeName = extractRelativePath(mProject->directory(), unit->fileName());
         FileType fileType = getFileType(RelativeName);
-        if (fileType==FileType::ASM && !compilerSet()->canAssemble())
-            continue;
 
         if (fileType == FileType::CSource || fileType == FileType::CppSource
-                || fileType == FileType::ASM || fileType==FileType::GAS) {
+                || fileType==FileType::GAS) {
             if (!mProject->options().objectOutput.isEmpty()) {
                 // ofile = C:\MyProgram\obj\main.o
                 QString fullObjFile = includeTrailingPathDelimiter(mProject->options().objectOutput)
@@ -235,8 +233,6 @@ void ProjectCompiler::writeMakeDefines(QFile &file)
         cppCompileArguments+= " -D__DEBUG__";
     }
 
-    if (compilerSet()->canAssemble())
-        writeln(file,"ASM      = " + extractFileName(compilerSet()->assembler()));
     writeln(file,"CPP      = " + extractFileName(compilerSet()->cppCompiler()));
     writeln(file,"CC       = " + extractFileName(compilerSet()->CCompiler()));
 #ifdef Q_OS_WIN
@@ -287,24 +283,6 @@ void ProjectCompiler::writeMakeDefines(QFile &file)
 #ifdef Q_OS_WIN
     writeln(file,"WINDRESFLAGS  = " + mProject->options().resourceCmd);
 #endif
-    if (compilerSet()->canAssemble()) {
-        QString asmFlags;
-#ifdef Q_OS_WIN
-        if (Settings::CompilerSets::isTarget64Bit(compilerSet()->target())) {
-            if (mProject->getCompileOption(CC_CMD_OPT_POINTER_SIZE)=="32")
-                asmFlags = "-f win32 " + mProject->options().assemblerArgs;
-            else
-                asmFlags = "-f win64 " + mProject->options().assemblerArgs;
-        } else {
-            asmFlags = "-f win32 " + mProject->options().assemblerArgs;
-        }
-#elif defined(Q_OS_LINUX)
-        asmFlags = "-f elf64";
-#elif defined(Q_OS_MACOS)
-        asmFlags = "-f macho64";
-#endif
-        writeln(file,"ASMFLAGS      = " + asmFlags);
-    }
 
     // This needs to be put in before the clean command.
     if (mProject->options().type == ProjectType::DynamicLib) {
@@ -384,7 +362,6 @@ void ProjectCompiler::writeMakeObjFilesRules(QFile &file)
         FileType fileType = getFileType(unit->fileName());
         // Only process source files
         if (fileType!=FileType::CSource && fileType!=FileType::CppSource
-                && fileType!=FileType::ASM
                 && fileType!=FileType::GAS)
             continue;
 
@@ -504,17 +481,12 @@ void ProjectCompiler::writeMakeObjFilesRules(QFile &file)
                 if (!mOnlyCheckSyntax) {
                     writeln(file, "\t$(CC) -c " + genMakePath1(shortFileName) + " -o \"" + objFileName2 + "\" $(CFLAGS) " + encodingStr);
                 }
-            } else if (fileType==FileType::ASM) {
-                if (!mOnlyCheckSyntax) {
-                    writeln(file, "\t$(ASM) " + genMakePath1(shortFileName) + " -o \"" + objFileName2 + "\" $(ASMFLAGS) ");
-                }
             }
         }
     }
 
 #ifdef Q_OS_WIN
     if (!mProject->options().privateResource.isEmpty()) {
-
         // Concatenate all resource include directories
         QString ResIncludes(" ");
         for (int i=0;i<mProject->options().resourceIncludes.count();i++) {