+++ RedPandaIDE/mainwindow.h
@@ -251,7 +251,6 @@ public slots:
     void disableDebugActions();
     void enableDebugActions();
     void stopDebugForNoSymbolTable();
-    void stopDebugForNoSourceFile();
     void onTodoParsingFile(const QString& filename);
     void onTodoParseStarted();
     void onTodoFound(const QString& filename, int lineNo, int ch, const QString& line);