+++ RedPandaIDE/systemconsts.cpp
@@ -36,7 +36,6 @@ SystemConsts::SystemConsts(): mDefaultFileFilters()
     addDefaultFileFilter(QObject::tr("C++ files"),"*.cpp *.cc *.cxx");
     addDefaultFileFilter(QObject::tr("Header files"),"*.h *.hh *.hpp");
     addDefaultFileFilter(QObject::tr("GAS files"),"*.s *.S");
-    addDefaultFileFilter(QObject::tr("ASM files"),"*.asm");
     addDefaultFileFilter(QObject::tr("Lua files"),"*.lua");
 
     addFileFilter(mIconFileFilters, QObject::tr("Icon files"), "*.ico");